refactor: separate memo repository and user repository

+++ app/entity/memo.go
@@ -4,8 +4,8 @@ import "gorm.io/gorm"
 
 type Memo struct {
 	gorm.Model
-	UserID   uint64 `gorm:"foreignKey" json:"userId"`
-	BookID   uint64 `gorm:"foreignKey" json:"bookId"`
+	UserID   uint   `gorm:"foreignKey" json:"userId"`
+	BookID   uint   `gorm:"foreignKey" json:"bookId"`
 	Text     string `gorm:"column:text" json:"text"`
 	Category string `gorm:"column:category" json:"category"`
 }
